+++ shared/GR_Dataset.lua
@@ -4,7 +4,7 @@ ns.ds = {}
 local ds = ns.ds
 
 function ds:Init()
-    self.dbVersion = '3.0.0'
+    self.dbVersion = '3.0.1'
     self.tblRaces = self:races()
     self.tblClasses = self:classes()
     self.tblClassesByName = nil
@@ -19,8 +19,34 @@ function ds:WhatsNew() -- What's new in the current version
                     Discord: https://discord.gg/ZtS6Q2sKRH
                 (or click on the icon in the top left corner)|r
                 
-    |CFFFFFF00v3.0.0 Notes|r
-        - Updated for War Within
+    |CFFFFFF00v3.0 Notes|r
+        Alpha Fixes v3.0.6:
+        - Fixed preview showing |c2
+        - Fixed issue with lblPlayersScanned error.
+        - Right Click Invite Menu:
+            - Added option to send invite message.
+            - Now checks if player is in the guild.
+
+        I have done pretty much a full rewrite of Guild Recruiter.
+        I have added a lot of new features and fixed a lot of bugs.
+
+            ** CHANGES WERE MADE TO THE DATABASE. **
+        ** PLEASE BACKUP YOUR SAVE FILES BEFORE UPDATING. **
+
+        Features:
+        - Minimap icon:
+            - Added shift+click to directly open the scanner.
+            - Added anti-spam and black list counts to the tooltip.
+        - Invite Player Changes:
+            - Analytics now tracks /ginvite.
+            - Right Click Invite Menu:
+                - Invite to guild now works.
+                - Hides the option to black list if on the list.
+                - Ask if you try to invite someone on the black list.
+                - Send your invite message to a player.
+        - Scanner Changes:
+            - Whispers will now show when show whispers is enabled.
+            - Invite list and Who will show after leaving the screen.
     ]]
 
     return title, msg, height