+++ CHANGELOG.md
@@ -1,9 +1,30 @@
 # Guild Recruiter
-## 3.0.0 - War Within Release
+## 3.0 - War Within Release (Alpha)
+### Alpha Fixes v3.0:
+- Fixed preview showing |c2
+- Fixed issue with lblPlayersScanned error.
+- Right Click Invite Menu:
+    - Added option to send invite message.
+    - Now checks if player is in the guild.
+
 I have done pretty much a full rewrite of Guild Recruiter.  I have added a lot of new features and fixed a lot of bugs.
-Features:
-- Reworked Settings and the GM/Personal invite system should be more clear.
-- Black List will now immediately remove a player from the black list and there will not be a wait time.
-- Added the ability to add zones to the zone list.  This will require a bit of searching on the internet to find the zone ID.
-- Added Shift+Click to directly open scanner.
+
+    ** CHANGES WERE MADE TO THE DATABASE. **
+** PLEASE BACKUP YOUR SAVED FILES BEFORE UPDATING. **
+
+### Features:
+- Minimap icon:
+    - Added shift+click to directly open the scanner.
+    - Added anti-spam and black list counts to the tooltip.
+- Invite Player Changes:
+    - Analytics now tracks /ginvite.
+    - Right Click Invite Menu:
+        - Invite to guild now works.
+        - Hides the option to black list if on the list.
+        - Ask if you try to invite someone on the black list.
+        - Send your invite message to a player.
+- Scanner Changes:
+    - Whispers will now show when show whispers is enabled.
+    - Invite list and Who will show after leaving the screen.
+